\name{jgr.internals}
\alias{.completeCommand}
\alias{.generate.mac.script}
\alias{.generate.run.script}
\alias{.generate.windows.script}
\alias{.getContent}
\alias{.getDataObjects}
\alias{.getFunctionsInWS}
\alias{.getModels}
\alias{.getOtherObjects}
\alias{.jgr.pkg.path}
\alias{.jgr.register.function}
\alias{.jgr.save.JavaGD.as}
\alias{.jgr.works}
\alias{.refresh}
\alias{.refreshKeyWords}
\alias{.refreshObjects}
\title{
  Internal JGR functions
}
\description{
  The following functions are publicly visible, but with the expection
  of \code{JGR} they should be only invoked from the GUI and not directly
  by the user.
}
\keyword{programming}
